+++ web/package.json
@@ -23,7 +23,6 @@
         "lodash": "^4.17.21",
         "prop-types": "^15.8.1",
         "react": "^17.0.2",
-        "react-bootstrap": "^2.9.1",
         "react-dom": "^17.0.2",
         "react-grid-layout": "^1.3.4",
         "react-intersection-observer": "^9.5.2",
